moving afdesign specific options from shared model
