+++ colabdesign/af/utils.py
@@ -14,6 +14,35 @@ from colabdesign.af.alphafold.common import protein
 # AF_UTILS - various utils (save, plot, etc)
 ####################################################
 class _af_utils:  
+
+  def set_opt(self, *args, **kwargs):
+    '''
+    set [opt]ions
+    -------------------
+    note: model.restart() resets the [opt]ions to their defaults
+    use model.set_opt(..., set_defaults=True) 
+    or model.restart(..., reset_opt=False) to avoid this
+    -------------------    
+    model.set_opt(models=1, recycles=0)
+    model.set_opt(con=dict(num=1)) or set_opt({"con":{"num":1}})
+    model.set_opt(lr=1, set_defaults=True)
+    '''
+    if "best_metric" in kwargs:
+      self._args["best_metric"] = kwargs.pop("best_metric")
+
+    if "optimizer" in kwargs:
+      print("ERROR: use model.restart(optimizer=...) to set the optimizer")
+
+    if "recycle_mode" in kwargs:
+      if kwargs["recycle_mode"] in ["sample","last"] and self._args["recycle_mode"] in ["sample","last"]:
+        self._args["recycle_mode"] = kwargs.pop("recycle_mode")
+      else:
+        print(f"ERROR: use {self.__class__.__name__}(recycle_mode=...) to set the recycle_mode")
+
+    if kwargs.pop("set_defaults", False):
+      update_dict(self._opt, *args, **kwargs)
+
+    update_dict(self.opt, *args, **kwargs)
   
   def get_loss(self, x="loss"):
     '''output the loss (for entire trajectory)'''
